pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "remem"
version = "1.0.0"
description = "A sophisticated Agent Memory API powered by Redis Vectorset"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Rowan Trollope", email = "rowantrollope@gmail.com"}
]
maintainers = [
    {name = "Rowan Trollope", email = "rowantrollope@gmail.com"}
]
keywords = ["memory", "ai", "agent", "redis", "vectorset", "langchain", "langgraph"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.0.60",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "redis>=5.0.0",
    "numpy>=1.24.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.4.0",
    "mcp>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
docker = [
    "gunicorn>=21.0.0",
]

[project.urls]
Homepage = "https://github.com/rowantrollope/remem"
Repository = "https://github.com/rowantrollope/remem"
Issues = "https://github.com/rowantrollope/remem/issues"
Documentation = "https://github.com/rowantrollope/remem#readme"

[project.scripts]
remem = "main:main"
remem-web = "web_app:main"
remem-mcp = "mcp_server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["memory*", "api*", "clients*", "llm*"]
exclude = ["tests*", "docs*", "scripts*"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "redis.*",
    "openai.*",
    "langchain.*",
    "langgraph.*",
    "mcp.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
